added cool cards to dashboard
