+++ excel-tracker-application/src/components/Dashboard.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 
 import Colors from './../assets/text/Colors';
 import Card from 'react-bootstrap/Card';
+import CoolCard from './coolcard.js'
 
 import logo from './../assets/images/logo.svg';
 import competencies from './../assets/images/competencies.png'
@@ -25,17 +26,7 @@ class Dashboard extends React.Component {
                 Hello, {this.props.currentUser}
               </a>
               <br></br>
-              <Card style={{ width: '18rem' }}>
-                <Card.Img variant="top" src={competencies} />
-                <Card.Body>
-                  <Card.Title>Competencies</Card.Title>
-                  <Card.Text>
-                    Some quick example text to build on the card title and make up the bulk of
-                    the card's content.
-                  </Card.Text>
-                  <Button variant="primary">Go somewhere</Button>
-                </Card.Body>
-              </Card>
+              <CoolCard />
               <br></br>
             </header>
           </div>
